Tabs Styles turned into global
